total reworking of list (dynamic array)
