+++ src/core/dm_mem.h
@@ -12,6 +12,7 @@ typedef enum dm_mem_tag
 	DM_MEM_LIST,
 	DM_MEM_MAP,
 	DM_MEM_RENDERER,
+	DM_MEM_RENDER_COMMAND,
 	DM_MEM_RENDER_PIPELINE,
 	DM_MEM_RENDERER_BUFFER,
 	DM_MEM_RENDERER_SHADER,
