+++ src/structures/dm_old_map.h
@@ -1,0 +1,36 @@
+#ifndef __DM_MAP_H__
+#define __DM_MAP_H__
+
+#define CAPACITY 50000
+
+typedef struct dm_map_item
+{
+    char* key;
+    char* value;
+} dm_map_item;
+
+typedef struct dm_map_link_list
+{
+    dm_map_item* item;
+    struct dm_map_link_list* next;
+} dm_map_link_list;
+
+typedef struct dm_map
+{
+    int size, count;
+    dm_map_item** items;
+    dm_map_link_list** overflows;
+} dm_map;
+
+unsigned long simple_hash_function(const char* str);
+dm_map* dm_map_create(int size);
+dm_map_item* dm_map_create_item(char* key, char* value);
+void dm_map_delete_item(dm_map_item* item);
+void dm_map_delete(dm_map* map);
+
+void dm_map_insert(dm_map* map, char* key, char* value);
+char* dm_map_search(dm_map* map, char* key);
+
+void dm_map_print(dm_map* map);
+void dm_map_search_print(dm_map* map, char* key);
+#endif
